ada: Fix restoration of parent link
When resetting the parent link after having restored the selected
component node, the assertion used was incorrectly triggered when the
traversal hits the members of the parameters association list, as in:

   This.Some_Func (Param1, Param2).Dispatching_Call

When restoring the selected component for Dispatching_Call, the
assertion was wrongly triggered when passed Param1 and Param2.

gcc/ada/

	* contracts.adb (Restore_Original_Selected_Component): Adjust assertion.

Tested on x86_64-pc-linux-gnu, committed on master.