ada: Fix retrieval of spec entity from entry body entity
When retrieving entities of subprogram spec we only handled functions
and procedures, but not entries. This had no consequences, because we
then only applied checks to functions, but still is worth a cleanup, so
the code is easier to adapt for access-to-subprogram entities as well.

gcc/ada/

	* sem_util.adb (Check_Result_And_Post_State): Properly handle entry
	bodies.

Tested on x86_64-pc-linux-gnu, committed on master.