ada: Use Code_Address attribute to determine subprogram addresses
The runtime used label addresses to determine the code address of
subprograms because the subprogram's canonical address on some targets
is a descriptor or a stub. Simplify the code by using the Code_Address
attribute instead, which is designed to return the code address of a
subprogram. This also works around a current GNAT-LLVM limitation where
the address of a label is incorrectly calculated when using -O1. As a
result, we can now build a-except.adb and g-debpoo.adb at -O1 again with
GNAT-LLVM.

gcc/ada/

	* libgnat/a-excach.adb (Call_Chain): Replace
	Code_Address_For_AAA/ZZZ functions with AAA/ZZZ'Code_Address.
	* libgnat/a-except.adb (Code_Address_For_AAA/ZZZ): Delete.
	(AAA/ZZZ): New null procedures.
	* libgnat/g-debpoo.adb
	(Code_Address_For_Allocate_End): Delete.
	(Code_Address_For_Deallocate_End): Delete.
	(Code_Address_For_Dereference_End): Delete.
	(Allocate): Remove label and use Code_Address attribute to
	determine subprogram addresses.
	(Dellocate): Likewise.
	(Dereference): Likewise.
	(Allocate_End): Convert to null procedure.
	(Dellocate_End): Likewise.
	(Dereference_End): Likewise.

Tested on x86_64-pc-linux-gnu, committed on master.