ada: Fix incorrect copies being used with 'Address
When using 'Address on an object with a size clause, gigi would end up
creating a copy and using its address instead of the one of the original
object, leading to incorrect behavior. Remove the conversion (that
triggers the copy) when 'Address is applied to a declaration.

gcc/ada/

	* gcc-interface/trans.cc (Attribute_to_gnu): Also strip conversion
	in case of DECL.

Tested on x86_64-pc-linux-gnu, committed on master.