doc: Fix description of x86 -m32 option [PR109954]
In https://gcc.gnu.org/PR109954 I suggested also adding:

"N.B., using @option{-march} might be required to produce code suitable
for a specific CPU family, e.g., @option{-march=i486}."

I realise that that is true for all of -m32, -m64 and -mx32, and similar
rules apply for other targets too. But I still feel that saying it
explicitly for -m32 doesn't hurt, and would avoid a common
misunderstanding by putting that info somewhere it's more likely to be
read.

But I'd prefer to just fix the part that is *wrong*, and then we can
discuss whether or not that other part is an improvement. This patch
fixes the wrongness.

OK for trunk and release branches?

-- >8 --

This option does not imply -march=i386 so it's incorrect to say it
generates code that will run on "any i386 system".

gcc/ChangeLog:

	PR target/109954
	* doc/invoke.texi (x86 Options): Fix description of -m32 option.