c++: refine dependent_alias_template_spec_p [PR90679]
For a complex alias template-id, dependent_alias_template_spec_p returns
true if any template argument of the template-id is dependent.  This
predicate indicates that substitution into the template-id may behave
differently with respect to SFINAE than substitution into the expanded
alias, and so the alias is in a way non-transparent.  For example
'first_t<T, T&>' in

  template<class T, class...> using first_t = T;
  template<class T> first_t<T, T&> f();

is such an alias template-id since first_t doesn't use its second
template parameter and so the substitution into the expanded alias would
discard the SFINAE effects of the corresponding (dependent) argument 'T&'.

But this predicate is overly conservative since what really matters for
sake of SFINAE equivalence is whether a template argument corresponding
to an _unused_ template parameter is dependent.  So the predicate should
return false for e.g. 'first_t<T, int>' or 'first_t<T>'.

This patch refines the predicate appropriately.  We need to be able to
efficiently determine which template parameters of a complex alias
template are unused, so to that end we add a new out parameter to
complex_alias_template_p and cache its result in an on-the-side
hash_map that replaces the existing TEMPLATE_DECL_COMPLEX_ALIAS_P
flag.  And in doing so, we fix a latent bug that this flag wasn't
being propagated during partial instantiation, and so we were treating
all partially instantiated member alias templates as non-complex.

	PR c++/90679

gcc/cp/ChangeLog:

	* cp-tree.h (TEMPLATE_DECL_COMPLEX_ALIAS_P): Remove.
	(most_general_template): Constify parameter.
	* pt.cc (push_template_decl): Adjust after removing
	TEMPLATE_DECL_COMPLEX_ALIAS_P.
	(complex_alias_tmpl_info): New hash_map.
	(uses_all_template_parms_data::seen): Change type to
	tree* from bool*.
	(complex_alias_template_r): Adjust accordingly.
	(complex_alias_template_p): Add 'seen_out' out parameter.
	Call most_general_template and check PRIMARY_TEMPLATE_P.
	Use complex_alias_tmpl_info to cache the result and set
	'*seen_out' accordigly.
	(dependent_alias_template_spec_p): Add !processing_template_decl
	early exit test.  Consider dependence of only template arguments
	corresponding to seen template parameters as per

gcc/testsuite/ChangeLog:

	* g++.dg/cpp0x/alias-decl-75.C: New test.