rs6000: Fix expected counts powerpc/p9-vec-length-full
GCC maintainers:

The following patch updates the expected instruction counts in four
tests.  The counts in all of the tests changed with commit
f574e2dfae79055f16d0c63cc12df24815d8ead6.

The updated counts have been verified on both Power 9 and Power 10.

Please let me know if this patch is acceptable for mainline.  Thanks.

                      Carl

--------------------
rs6000: Fix expected counts powerpc/p9-vec-length-full tests

The counts for instructions lxvl and stxvl in tests:

  p9-vec-length-full-1.c
  p9-vec-length-full-2.c
  p9-vec-length-full-6.c
  p9-vec-length-full-7.c

changed with commit:

   commit f574e2dfae79055f16d0c63cc12df24815d8ead6
   Author: Ju-Zhe Zhong <juzhe.zhong@rivai.ai>
   Date:   Thu May 25 22:42:35 2023 +0800

     VECT: Add decrement IV iteration loop control by variable amount support

     This patch is supporting decrement IV by following the flow designed by
     Richard:
           ...

The expected counts for lxvl changed from 20 to 40 and the counts for stxvl
changed from 10 to 20 in the first three tests.  The number of stxvl
instructions changed from 12 to 20 in p9-vec-length-full-7.c.  This
patch updates the number of expected instructions in the four tests.

The counts have been verified on Power 9 and Power 10.